# Speed up build-cli-usage.sh script via adding background processes